refactor(cmd): fold audit_meta_schema spike into audit_tokens --compare-schemas
audit_meta_schema was a sizing spike (no Makefile target, no CI gate) that
duplicated audit_tokens' bytes/4 heuristic and in-memory MCP setup. Its logic
now lives behind audit_tokens --compare-schemas, producing byte-identical output
to the former standalone binary (verified against plan/baselines/phase0/).

The repeat() helper is replaced by strings.Repeat (already imported) and
human() is renamed to humanBytes() to avoid ambiguity with the token audit's
own formatting helpers. net cmd/ count: 23 → 22.
