+++ docs/development/tool-surfaces-and-action-core.md
@@ -378,8 +378,9 @@ The migration is enforced by source-level tests and audits:
 - `TestActionCatalog_BaselineCountsDoNotRegress` keeps CE, Enterprise, and
   GitLab.com Enterprise catalog action counts stable.
 - `make audit-dynamic-aliases`, `go run ./cmd/audit_output/`,
-  `go run ./cmd/audit_tools/`, `go run ./cmd/audit_meta_schema/`, and
-  `go run ./cmd/audit_tokens/` are the post-metadata-change surface audits.
+  `go run ./cmd/audit_tools/`, and
+  `go run ./cmd/audit_tokens/` (with `--compare-schemas` for the meta-tool
+  InputSchema sizing spike) are the post-metadata-change surface audits.
 
 ## Useful Checks
